@@ -15,67 +15,82 @@
     "higher order component",
     "hoc"
   ],
-  "main": "dist/formik.js",
-  "module": "dist/formik.es6.js",
-  "typings": "dist/formik.d.ts",
+  "main": "dist/index.js",
+  "module": "dist/formik.esm.js",
+  "typings": "dist/index.d.ts",
   "files": [
     "dist"
   ],
-  "scripts": {
-    "test": "jest --env=jsdom",
-    "test:watch": "npm run test -- --watch",
-    "start": "cross-env NODE_ENV=development tsc-watch --onSuccess \"rollup -c\"",
-    "prebuild": "rimraf dist",
-    "build": "tsc && cross-env NODE_ENV=production rollup -c && cross-env NODE_ENV=development rollup -c && rimraf compiled",
-    "prepublish": "npm run build",
-    "format": "prettier --trailing-comma es5 --single-quote --write 'src/**/*' 'test/**/*' 'README.md'",
-    "precommit": "lint-staged",
-    "addc": "all-contributors add",
-    "gen-docs": "all-contributors generate && doctoc README.md"
-  },
-  "dependencies": {
-    "lodash.isequal": "4.5.0",
-    "prop-types": "^15.5.10",
-    "warning": "^3.0.0"
-  },
   "peerDependencies": {
     "react": ">=15"
   },
+  "scripts": {
+    "test": "jest --env=jsdom",
+    "test:watch": "npm run test -- --watch",
+    "start": "cp ./index.js ./dist/index.js && cross-env NODE_ENV=development rollup -w -c | tsc -w -p tsconfig.base.json",
+    "prebuild": "rimraf dist",
+    "build": "cross-env NODE_ENV=production tsc -p tsconfig.base.json  && rollup -c && rimraf compiled && cp ./index.js ./dist/index.js",
+    "prepublish": "npm run build",
+    "format": "prettier --trailing-comma es5 --single-quote --write 'src/**/*' 'test/**/*' 'README.md'",
+    "precommit": "lint-staged",
+    "storybook": "start-storybook -p 9001",
+    "addc": "all-contributors add",
+    "start-website": "cd website && yarn start",
+    "build-website": "cd website && yarn install && yarn build",
+    "gen-docs": "all-contributors generate && doctoc README.md",
+    "size": "size-limit"
+  },
+  "dependencies": {
+    "create-react-context": "^0.2.2",
+    "deepmerge": "^2.1.1",
+    "hoist-non-react-statics": "^2.5.5",
+    "lodash.clonedeep": "^4.5.0",
+    "lodash.topath": "4.5.2",
+    "prop-types": "^15.6.1",
+    "react-fast-compare": "^1.0.0",
+    "tslib": "^1.9.3",
+    "warning": "^3.0.0"
+  },
   "optionalDependencies": {},
   "devDependencies": {
-    "@types/enzyme": "2.8.4",
-    "@types/jest": "20.0.6",
-    "@types/lodash.isequal": "4.5.2",
-    "@types/node": "8.0.19",
-    "@types/prop-types": "15.5.1",
-    "@types/react": "16.0.0",
-    "@types/react-dom": "15.5.1",
+    "@pisano/enzyme": "^3.3.0-pisano.public.1",
+    "@pisano/enzyme-adapter-react-16": "^1.1.1-pisano.public.1",
+    "@storybook/addon-options": "^3.4.8",
+    "@storybook/react": "^3.4.0",
+    "@types/jest": "^22.2.3",
+    "@types/lodash.clonedeep": "^4.5.3",
+    "@types/lodash.topath": "4.5.3",
+    "@types/react": "16.0.28",
+    "@types/react-dom": "^16.0.3",
     "@types/react-test-renderer": "15.5.2",
     "@types/warning": "^3.0.0",
+    "@types/yup": "^0.24.9",
     "all-contributors-cli": "^4.4.0",
+    "awesome-typescript-loader": "^3.4.1",
     "cross-env": "5.0.5",
     "doctoc": "^1.3.0",
-    "enzyme": "2.9.1",
     "husky": "0.14.3",
-    "jest": "20.0.4",
+    "jest": "^23.1.0",
+    "jest-cli": "^23.1.0",
     "lint-staged": "4.0.2",
-    "prettier": "^1.8.2",
-    "react": "15.6.1",
-    "react-dom": "15.6.1",
-    "react-test-renderer": "15.6.1",
-    "rimraf": "2.6.1",
-    "rollup": "0.45.2",
-    "rollup-plugin-commonjs": "8.1.0",
+    "prettier": "1.11.1",
+    "raw-loader": "^0.5.1",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "rimraf": "^2.6.2",
+    "rollup": "^0.59.0",
+    "rollup-plugin-commonjs": "^9.1.3",
     "rollup-plugin-filesize": "1.4.2",
-    "rollup-plugin-node-resolve": "3.0.0",
-    "rollup-plugin-replace": "1.1.1",
-    "rollup-plugin-sourcemaps": "0.4.2",
-    "rollup-plugin-uglify": "2.0.1",
-    "ts-jest": "20.0.9",
-    "tsc-watch": "1.0.7",
+    "rollup-plugin-node-resolve": "^3.3.0",
+    "rollup-plugin-replace": "^2.0.0",
+    "rollup-plugin-sourcemaps": "^0.4.2",
+    "rollup-plugin-uglify": "^3.0.0",
+    "size-limit": "^0.17.0",
+    "ts-jest": "^22.4.6",
+    "tsc-watch": "^1.0.21",
     "tslint": "5.5.0",
     "tslint-react": "3.2.0",
-    "typescript": "2.4.2",
+    "typescript": "^3.0.1",
     "yup": "0.21.3"
   },
   "lint-staged": {
