@@ -1,22 +1,11 @@
 import * as React from 'react';
-import * as renderer from 'react-test-renderer';
+import * as ReactDOM from 'react-dom';
+import * as Yup from 'yup';
+import { Formik, FormikProps } from '../src';
+import { shallow, mount } from '@pisano/enzyme';
+import { sleep, noop } from './testHelpers';
 
-import { Formik, FormikConfig, FormikProps } from '../src/formik';
-import { mount, shallow } from 'enzyme';
-
-const Yup = require('yup');
-
-// tslint:disable-next-line:no-empty
-const noop = () => {};
-
-const sleep = ms => new Promise(resolve => setTimeout(resolve, ms));
-
-interface Props {
-  user: {
-    name: string;
-  };
-  someFunction?: () => void;
-}
+jest.spyOn(global.console, 'error');
 
 interface Values {
   name: string;
@@ -24,10 +13,10 @@ interface Values {
 
 const Form: React.SFC<FormikProps<Values>> = ({
   values,
+  touched,
   handleSubmit,
   handleChange,
   handleBlur,
-  setStatus,
   status,
   errors,
   isSubmitting,
@@ -41,14 +30,8 @@ const Form: React.SFC<FormikProps<Values>> = ({
         value={values.name}
         name="name"
       />
-      {errors.name && <div id="feedback">{errors.name}</div>}
+      {touched.name && errors.name && <div id="feedback">{errors.name}</div>}
       {isSubmitting && <div id="submitting">Submitting</div>}
-      <button
-        id="statusButton"
-        onClick={() => setStatus({ myStatusMessage: 'True' })}
-      >
-        Call setStatus
-      </button>
       {status &&
         !!status.myStatusMessage && (
           <div id="statusMessage">{status.myStatusMessage}</div>
@@ -62,7 +45,26 @@ const BasicForm = (
   <Formik initialValues={{ name: 'jared' }} onSubmit={noop} component={Form} />
 );
 
-describe('Formik Next', () => {
+class WithState extends React.Component<{}, { data: { name: string } }> {
+  constructor(props: {}) {
+    super(props);
+    this.state = {
+      data: { name: 'ivan' },
+    };
+  }
+
+  render() {
+    return (
+      <Formik
+        initialValues={this.state.data}
+        onSubmit={noop}
+        component={Form}
+      />
+    );
+  }
+}
+
+describe('<Formik>', () => {
   it('should initialize Formik state and pass down props', () => {
     const tree = shallow(BasicForm);
     expect(tree.find(Form).props().isSubmitting).toBe(false);
@@ -71,6 +73,7 @@ describe('Formik Next', () => {
     expect(tree.find(Form).props().errors).toEqual({});
     expect(tree.find(Form).props().dirty).toBe(false);
     expect(tree.find(Form).props().isValid).toBe(false);
+    expect(tree.find(Form).props().submitCount).toBe(0);
   });
 
   describe('FormikHandlers', () => {
@@ -259,6 +262,42 @@ describe('Formik Next', () => {
         expect(preventDefault).toHaveBeenCalled();
       });
 
+      it('should not error if called without an object', () => {
+        const FormNoEvent = (
+          <Formik initialValues={{ name: 'jared' }} onSubmit={noop}>
+            {({ handleSubmit }) => (
+              <button
+                onClick={() =>
+                  handleSubmit(undefined as any /* undefined event */)
+                }
+              />
+            )}
+          </Formik>
+        );
+        const tree = mount(FormNoEvent);
+        const fn = () => {
+          tree.find('button').simulate('click');
+        };
+        expect(fn).not.toThrow();
+      });
+
+      it('should not error if called without preventDefault property', () => {
+        const FormNoPreventDefault = (
+          <Formik initialValues={{ name: 'jared' }} onSubmit={noop}>
+            {({ handleSubmit }) => (
+              <button
+                onClick={() => handleSubmit({} as any /* no preventDefault */)}
+              />
+            )}
+          </Formik>
+        );
+        const tree = mount(FormNoPreventDefault);
+        const fn = () => {
+          tree.find('button').simulate('click');
+        };
+        expect(fn).not.toThrow();
+      });
+
       it('should touch all fields', () => {
         const tree = shallow(BasicForm);
         tree
@@ -291,32 +330,15 @@ describe('Formik Next', () => {
 
         expect(
           tree
+            .update()
             .find(Form)
             .dive()
             .find('#submitting')
         ).toHaveLength(1);
       });
 
-      // it('should correctly map form values to payload', () => {
-      //   interface Payload {
-      //     user: { name: string };
-      //   }
-      //   const CustomPayloadForm = Formik<Props, Values, Payload>({
-      //     mapPropsToValues: ({ user }) => ({ ...user }),
-      //     mapValuesToPayload: ({ name }) => ({ user: { name } }),
-      //     onSubmit: payload => {
-      //       expect(payload).toEqual({ user: { name: 'jared' } });
-      //       expect(payload).not.toEqual({ name: 'jared' });
-      //     },
-      //   })(Form);
-      //   const tree = shallow(<CustomPayloadForm user={{ name: 'jared' }} />);
-      //   tree.find(Form).dive().find('form').simulate('submit', {
-      //     preventDefault: noop,
-      //   });
-      // });
-
       describe('with validate (SYNC)', () => {
-        it('should call validate if present', () => {
+        it('should call validate if present', async () => {
           const validate = jest.fn().mockReturnValue({});
           const tree = shallow(
             <Formik
@@ -326,17 +348,14 @@ describe('Formik Next', () => {
               validate={validate}
             />
           );
-          tree
+          await tree
             .find(Form)
-            .dive()
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
           expect(validate).toHaveBeenCalled();
         });
 
-        it('should submit the form if valid', () => {
+        it('should submit the form if valid', async () => {
           const onSubmit = jest.fn();
           const tree = shallow(
             <Formik
@@ -346,17 +365,14 @@ describe('Formik Next', () => {
               validate={noop}
             />
           );
-          tree
+          await tree
             .find(Form)
-            .dive()
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
           expect(onSubmit).toHaveBeenCalled();
         });
 
-        it('should not submit the form if invalid', () => {
+        it('should not submit the form if invalid', async () => {
           const validate = jest.fn().mockReturnValue({ name: 'Error!' });
           const onSubmit = jest.fn();
 
@@ -368,20 +384,18 @@ describe('Formik Next', () => {
               validate={validate}
             />
           );
-          tree
+          await tree
             .find(Form)
-            .dive()
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
+
           expect(validate).toHaveBeenCalled();
           expect(onSubmit).not.toHaveBeenCalled();
         });
       });
 
       describe('with validate (ASYNC)', () => {
-        it('should call validate if present', () => {
+        it('should call validate if present', async () => {
           const validate = jest.fn(() => Promise.resolve({}));
 
           const tree = shallow(
@@ -392,13 +406,10 @@ describe('Formik Next', () => {
               validate={validate}
             />
           );
-          tree
+          await tree
             .find(Form)
-            .dive()
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
           expect(validate).toHaveBeenCalled();
         });
 
@@ -446,7 +457,7 @@ describe('Formik Next', () => {
       });
 
       describe('with validationSchema (ASYNC)', () => {
-        it('should run validationSchema if present', () => {
+        it('should run validationSchema if present', async () => {
           const validate = jest.fn(() => Promise.resolve({}));
           const tree = shallow(
             <Formik
@@ -459,17 +470,16 @@ describe('Formik Next', () => {
               }}
             />
           );
-          tree
+
+          await tree
             .find(Form)
-            .dive()
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
+
           expect(validate).toHaveBeenCalled();
         });
 
-        it('should call validationSchema if it is a function and present', () => {
+        it('should call validationSchema if it is a function and present', async () => {
           const validate = jest.fn(() => Promise.resolve({}));
           const tree = shallow(
             <Formik
@@ -482,13 +492,12 @@ describe('Formik Next', () => {
               })}
             />
           );
-          tree
+
+          await tree
             .find(Form)
-            .dive()
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
+
           expect(validate).toHaveBeenCalled();
         });
       });
@@ -504,10 +513,9 @@ describe('Formik Next', () => {
         .setValues({ name: 'ian' });
       expect(
         tree
+          .update()
           .find(Form)
-          .dive()
-          .find('input')
-          .props().value
+          .props().values.name
       ).toEqual('ian');
     });
 
@@ -555,10 +563,9 @@ describe('Formik Next', () => {
         .setFieldValue('name', 'ian');
       expect(
         tree
+          .update()
           .find(Form)
-          .dive()
-          .find('input')
-          .props().value
+          .props().values.name
       ).toEqual('ian');
     });
 
@@ -606,7 +613,12 @@ describe('Formik Next', () => {
         .find(Form)
         .props()
         .setTouched({ name: true });
-      expect(tree.find(Form).props().touched).toEqual({ name: true });
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().touched
+      ).toEqual({ name: true });
     });
 
     it('setTouched should NOT run validations by default', async () => {
@@ -651,14 +663,34 @@ describe('Formik Next', () => {
         .find(Form)
         .props()
         .setFieldTouched('name', true);
-      expect(tree.find(Form).props().touched).toEqual({ name: true });
-      expect(tree.find(Form).props().dirty).toBe(true);
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().touched
+      ).toEqual({ name: true });
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().dirty
+      ).toBe(false);
       tree
         .find(Form)
         .props()
         .setFieldTouched('name', false);
-      expect(tree.find(Form).props().touched).toEqual({ name: false });
-      expect(tree.find(Form).props().dirty).toBe(false);
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().touched
+      ).toEqual({ name: false });
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().dirty
+      ).toBe(false);
     });
 
     it('setFieldTouched should run validations when validateOnBlur is true', async () => {
@@ -704,12 +736,12 @@ describe('Formik Next', () => {
         .find(Form)
         .props()
         .setErrors({ name: 'Required' });
+
       expect(
         tree
+          .update()
           .find(Form)
-          .dive()
-          .find('#feedback')
-          .text()
+          .props().errors.name
       ).toEqual('Required');
     });
 
@@ -721,26 +753,26 @@ describe('Formik Next', () => {
         .setFieldError('name', 'Required');
       expect(
         tree
+          .update()
           .find(Form)
-          .dive()
-          .find('#feedback')
-          .text()
+          .props().errors.name
       ).toEqual('Required');
     });
 
     it('setStatus sets status object', async () => {
       const tree = shallow(BasicForm);
+      const status = 'status';
       tree
         .find(Form)
-        .dive()
-        .find('#statusButton')
-        .simulate('click');
+        .props()
+        .setStatus(status);
+
       expect(
         tree
+          .update()
           .find(Form)
-          .dive()
-          .find('#statusMessage')
-      ).toHaveLength(1);
+          .props().status
+      ).toEqual(status);
     });
   });
 
@@ -748,7 +780,7 @@ describe('Formik Next', () => {
     it('should compute dirty as soon as any input is touched', () => {
       const tree = shallow(BasicForm);
       expect(tree.find(Form).props().dirty).toBe(false);
-      tree.setState({ touched: { name: true } });
+      tree.setState({ values: { name: 'ian' } });
       expect(tree.find(Form).props().dirty).toBe(true);
     });
 
@@ -758,7 +790,7 @@ describe('Formik Next', () => {
           initialValues={{ name: 'jared' }}
           onSubmit={noop}
           component={Form}
-          isInitialValid={props => true}
+          isInitialValid={() => true}
         />
       );
       expect(tree.find(Form).props().dirty).toBe(false);
@@ -771,7 +803,7 @@ describe('Formik Next', () => {
           initialValues={{ name: 'jared' }}
           onSubmit={noop}
           component={Form}
-          isInitialValid={props => false}
+          isInitialValid={() => false}
         />
       );
       expect(tree.find(Form).props().dirty).toBe(false);
@@ -806,21 +838,36 @@ describe('Formik Next', () => {
 
     it('should compute isValid if the form is dirty and there are errors', () => {
       const tree = shallow(BasicForm);
-      tree.setState({ touched: { name: true }, errors: { name: 'Required!' } });
+      tree.setState({ values: { name: 'ian' }, errors: { name: 'Required!' } });
       expect(tree.find(Form).props().dirty).toBe(true);
       expect(tree.find(Form).props().isValid).toBe(false);
     });
 
     it('should compute isValid if the form is dirty and there are not errors', () => {
       const tree = shallow(BasicForm);
-      tree.setState({ touched: { name: true } });
+      tree.setState({ values: { name: 'ian' } });
       expect(tree.find(Form).props().dirty).toBe(true);
       expect(tree.find(Form).props().isValid).toBe(true);
     });
+
+    it('should increase submitCount after submitting the form', async () => {
+      const tree = shallow(BasicForm);
+      expect(tree.find(Form).props().submitCount).toBe(0);
+      await tree
+        .find(Form)
+        .props()
+        .submitForm();
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().submitCount
+      ).toBe(1);
+    });
   });
 
-  describe('componentWillReceiveProps', () => {
-    let form, defaultForm, initialValues;
+  describe('componentDidUpdate', () => {
+    let form: any, initialValues: any;
     beforeEach(() => {
       initialValues = {
         name: 'formik',
@@ -837,7 +884,7 @@ describe('Formik Next', () => {
 
     it('should not resetForm if new initialValues are the same as previous', () => {
       const newInitialValues = Object.assign({}, initialValues);
-      form.componentWillReceiveProps({
+      form.componentDidUpdate({
         initialValues: newInitialValues,
         onSubmit: jest.fn(),
       });
@@ -849,7 +896,7 @@ describe('Formik Next', () => {
         ...initialValues,
         watchers: ['jared', 'ian', 'sam'],
       };
-      form.componentWillReceiveProps({
+      form.componentDidUpdate({
         initialValues: newInitialValues,
         onSubmit: jest.fn(),
       });
@@ -861,7 +908,7 @@ describe('Formik Next', () => {
         ...initialValues,
         github: { repoUrl: 'different' },
       };
-      form.componentWillReceiveProps({
+      form.componentDidUpdate({
         initialValues: newInitialValues,
         onSubmit: jest.fn(),
       });
@@ -878,11 +925,406 @@ describe('Formik Next', () => {
         ...initialValues,
         watchers: ['jared', 'ian', 'sam'],
       };
-      form.componentWillReceiveProps({
+      form.componentDidUpdate({
         initialValues: newInitialValues,
         onSubmit: jest.fn(),
       });
       expect(form.resetForm).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('handleReset', () => {
+    it('should call onReset with values and actions when form is reset', async () => {
+      const onReset = jest.fn();
+
+      const tree = shallow(
+        <Formik
+          initialValues={{ name: 'jared' }}
+          onSubmit={jest.fn()}
+          onReset={onReset}
+          component={Form}
+        />
+      );
+      await tree
+        .find(Form)
+        .props()
+        .handleReset();
+
+      expect(onReset).toHaveBeenCalledWith(
+        { name: 'jared' },
+        expect.objectContaining({
+          resetForm: expect.any(Function),
+          setError: expect.any(Function),
+          setErrors: expect.any(Function),
+          setFieldError: expect.any(Function),
+          setFieldTouched: expect.any(Function),
+          setFieldValue: expect.any(Function),
+          setStatus: expect.any(Function),
+          setSubmitting: expect.any(Function),
+          setTouched: expect.any(Function),
+          setValues: expect.any(Function),
+          submitForm: expect.any(Function),
+        })
+      );
+    });
+
+    it('should not error resetting form if onReset is not a prop', async () => {
+      const onSubmit = jest.fn();
+
+      const tree = shallow(
+        <Formik
+          initialValues={{ name: 'bar' }}
+          onSubmit={onSubmit}
+          component={Form}
+        />
+      );
+      await tree
+        .find(Form)
+        .props()
+        .handleReset();
+
+      expect(true);
+    });
+
+    it('should call onReset with values and actions when onReset is a promise', async () => {
+      const onReset = jest.fn(() => Promise.resolve('data'));
+
+      const tree = shallow(
+        <Formik
+          initialValues={{ name: 'jared' }}
+          onSubmit={jest.fn()}
+          onReset={onReset}
+          component={Form}
+        />
+      );
+
+      (tree.instance() as any).resetForm = jest.fn();
+
+      await tree
+        .find(Form)
+        .props()
+        .handleReset();
+
+      expect(onReset).toHaveBeenCalledWith(
+        { name: 'jared' },
+        expect.objectContaining({
+          resetForm: expect.any(Function),
+          setError: expect.any(Function),
+          setErrors: expect.any(Function),
+          setFieldError: expect.any(Function),
+          setFieldTouched: expect.any(Function),
+          setFieldValue: expect.any(Function),
+          setStatus: expect.any(Function),
+          setSubmitting: expect.any(Function),
+          setTouched: expect.any(Function),
+          setValues: expect.any(Function),
+          submitForm: expect.any(Function),
+        })
+      );
+
+      expect((tree.instance() as any).resetForm).toHaveBeenCalledWith('data');
+    });
+
+    it('should reset dirty flag even if initialValues has changed', async () => {
+      const tree = mount(<WithState />);
+      expect(tree.find(Form).props().dirty).toEqual(false);
+
+      tree
+        .find(Form)
+        .find('input')
+        .simulate('change', {
+          persist: noop,
+          target: {
+            id: 'name',
+            value: 'Ian',
+          },
+        });
+
+      expect(tree.find(Form).props().dirty).toEqual(true);
+
+      tree.setState({ data: { name: 'Jared' } });
+
+      await tree
+        .find(Form)
+        .props()
+        .handleReset();
+
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().dirty
+      ).toEqual(false);
+    });
+
+    it('should reset submitCount', () => {
+      const tree = mount(<WithState />);
+
+      tree
+        .find(Form)
+        .props()
+        .handleSubmit();
+
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().submitCount
+      ).toEqual(1);
+
+      tree
+        .find(Form)
+        .props()
+        .handleReset();
+
+      expect(
+        tree
+          .update()
+          .find(Form)
+          .props().submitCount
+      ).toEqual(0);
+    });
+  });
+
+  it('should warn against buttons with unspecified type', () => {
+    const FormWithNoButtonType = () => (
+      <Formik onSubmit={noop} initialValues={{ opensource: 'yay' }}>
+        {({ handleSubmit, handleChange, values }) => (
+          <form onSubmit={handleSubmit}>
+            <input
+              type="text"
+              onChange={handleChange}
+              value={values.opensource}
+              name="name"
+            />
+            <button>Submit</button>
+          </form>
+        )}
+      </Formik>
+    );
+    const tree = mount(<FormWithNoButtonType />);
+    const preventDefault = jest.fn();
+    const button = tree.find('button').getDOMNode();
+
+    button.focus(); // sets activeElement
+    tree.find('form').simulate('submit', {
+      preventDefault,
+    });
+    expect(global.console.error).toHaveBeenCalledWith(
+      expect.stringMatching(
+        /Warning: You submitted a Formik form using a button with an unspecified `type./
+      )
+    );
+
+    button.blur(); // unsets activeElement
+    (global.console.error as jest.Mock<{}>).mockClear();
+  });
+
+  it('should not warn when button has type submit', () => {
+    const FormWithValidButtonType = () => (
+      <Formik onSubmit={noop} initialValues={{ opensource: 'yay' }}>
+        {({ handleSubmit, handleChange, values }) => (
+          <form onSubmit={handleSubmit}>
+            <input
+              type="text"
+              onChange={handleChange}
+              value={values.opensource}
+              name="name"
+            />
+            <button type="submit">Submit</button>
+          </form>
+        )}
+      </Formik>
+    );
+    const tree = mount(<FormWithValidButtonType />);
+    const preventDefault = jest.fn();
+    const button = tree.find('button').getDOMNode();
+
+    button.focus(); // sets activeElement
+    tree.find('form').simulate('submit', {
+      preventDefault,
+    });
+    expect(global.console.error).not.toHaveBeenCalledWith(
+      expect.stringMatching(
+        /Warning: You submitted a Formik form using a button with an unspecified type./
+      )
+    );
+
+    button.blur(); // unsets activeElement
+    (global.console.error as jest.Mock<{}>).mockClear();
+  });
+
+  it('should not warn when activeElement is not a button', () => {
+    const FormWithInput = () => (
+      <Formik onSubmit={noop} initialValues={{ opensource: 'yay' }}>
+        {({ handleSubmit, handleChange, values }) => (
+          <form onSubmit={handleSubmit}>
+            <input
+              type="text"
+              onChange={handleChange}
+              value={values.opensource}
+              name="name"
+            />
+            <button type="submit">Submit</button>
+          </form>
+        )}
+      </Formik>
+    );
+    const tree = mount(<FormWithInput />);
+    const preventDefault = jest.fn();
+    const input = tree.find('input').getDOMNode();
+
+    input.focus(); // sets activeElement
+    tree.find('form').simulate('submit', {
+      preventDefault,
+    });
+    expect(global.console.error).not.toHaveBeenCalledWith(
+      expect.stringMatching(
+        /Warning: You submitted a Formik form using a button with an unspecified type./
+      )
+    );
+
+    input.blur(); // unsets activeElement
+    (global.console.error as jest.Mock<{}>).mockClear();
+  });
+  it('submit count increments', async () => {
+    const node = document.createElement('div');
+    const onSubmit = jest.fn();
+    let injected: any;
+    ReactDOM.render(
+      <Formik onSubmit={onSubmit} initialValues={{ opensource: 'yay' }}>
+        {formikProps => (injected = formikProps) && null}
+      </Formik>,
+      node
+    );
+
+    expect(injected.submitCount).toEqual(0);
+    await injected.submitForm();
+    expect(onSubmit).toHaveBeenCalled();
+    expect(injected.submitCount).toEqual(1);
+  });
+
+  it('isValidating is fired when submit is attempted', async () => {
+    const node = document.createElement('div');
+    const onSubmit = jest.fn();
+    const validate = jest.fn(() => ({ opensource: 'no ' }));
+    let injected: any;
+    ReactDOM.render(
+      <Formik
+        onSubmit={onSubmit}
+        validate={validate}
+        initialValues={{ opensource: 'yay' }}
+      >
+        {formikProps => (injected = formikProps) && null}
+      </Formik>,
+      node
+    );
+
+    expect(injected.submitCount).toEqual(0);
+    expect(injected.isSubmitting).toBe(false);
+    expect(injected.isValidating).toBe(false);
+    // we call set isValidating synchronously
+    const validatePromise = injected.submitForm();
+    // so it should change
+    expect(injected.isSubmitting).toBe(true);
+    expect(injected.isValidating).toBe(true);
+    // do it again async
+    await validatePromise;
+    // now both should be false because validation failed
+    expect(injected.isSubmitting).toBe(false);
+    expect(injected.isValidating).toBe(false);
+    expect(validate).toHaveBeenCalled();
+    expect(onSubmit).not.toHaveBeenCalled();
+    expect(injected.submitCount).toEqual(1);
+  });
+
+  it('isSubmitting is fired when submit is attempted', async () => {
+    const node = document.createElement('div');
+    const onSubmit = jest.fn();
+    const validate = jest.fn(() => ({}));
+    let injected: any;
+    ReactDOM.render(
+      <Formik
+        onSubmit={onSubmit}
+        validate={validate}
+        initialValues={{ opensource: 'yay' }}
+      >
+        {formikProps => (injected = formikProps) && null}
+      </Formik>,
+      node
+    );
+
+    expect(injected.submitCount).toEqual(0);
+    expect(injected.isSubmitting).toBe(false);
+    expect(injected.isValidating).toBe(false);
+    // we call set isValidating synchronously
+    const validatePromise = injected.submitForm();
+    // so it should change
+    expect(injected.isSubmitting).toBe(true);
+    expect(injected.isValidating).toBe(true);
+    // do it again async
+    await validatePromise;
+    // done validating
+    expect(injected.isValidating).toBe(false);
+    // now run submit
+    expect(injected.isSubmitting).toBe(true);
+    expect(validate).toHaveBeenCalled();
+    expect(onSubmit).toHaveBeenCalled();
+    expect(injected.submitCount).toEqual(1);
+  });
+
+  it('isValidating is fired validation is run', async () => {
+    const node = document.createElement('div');
+    const validate = jest.fn(() => ({ opensource: 'no' }));
+    let injected: any;
+    ReactDOM.render(
+      <Formik
+        onSubmit={noop}
+        validate={validate}
+        initialValues={{ opensource: 'yay' }}
+      >
+        {formikProps => (injected = formikProps) && null}
+      </Formik>,
+      node
+    );
+
+    expect(injected.isValidating).toBe(false);
+    // we call set isValidating synchronously
+    const validatePromise = injected.validateForm();
+    expect(injected.isValidating).toBe(true);
+    await validatePromise;
+    expect(validate).toHaveBeenCalled();
+    // so it should change
+    expect(injected.isValidating).toBe(false);
+  });
+
+  it('should merge validation errors', async () => {
+    const validate = () => ({
+      users: [{ firstName: 'required' }],
+    });
+    const validationSchema = Yup.object({
+      users: Yup.array().of(
+        Yup.object({
+          lastName: Yup.mixed().required('required'),
+        })
+      ),
+    });
+    let injected: any;
+
+    const tree = shallow(
+      <Formik
+        initialValues={{ users: [{ firstName: null, lastName: null }] }}
+        validate={validate}
+        validationSchema={validationSchema}
+        onSubmit={noop}
+      >
+        {formikProps => (injected = formikProps) && null}
+      </Formik>
+    );
+
+    await injected.validateForm();
+    expect(tree.state('errors')).toEqual({
+      users: [{ firstName: 'required', lastName: 'required' }],
     });
   });
 });
