@@ -1,17 +1,12 @@
 import * as React from 'react';
-import * as renderer from 'react-test-renderer';
 
-import { withFormik, InjectedFormikProps } from '../src/withFormik';
-import { mount, shallow } from 'enzyme';
-import { values } from '../src/utils';
-import { Formik } from '../src/formik';
-
-const Yup = require('yup');
+import { withFormik, FormikProps } from '../src';
+import { mount, shallow } from '@pisano/enzyme';
 
 // tslint:disable-next-line:no-empty
 const noop = () => {};
 
-const sleep = ms => new Promise(resolve => setTimeout(resolve, ms));
+const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
 
 interface Props {
   user: {
@@ -24,7 +19,7 @@ interface Values {
   name: string;
 }
 
-const Form: React.SFC<InjectedFormikProps<Props, Values>> = ({
+const Form: React.SFC<Props & FormikProps<Values>> = ({
   values,
   handleSubmit,
   handleChange,
@@ -79,6 +74,11 @@ describe('withFormik()', () => {
     expect(tree.find(Form).props().errors).toEqual({});
     expect(tree.find(Form).props().dirty).toBe(false);
     expect(tree.find(Form).props().isValid).toBe(false);
+  });
+
+  it('should correctly set displayName', () => {
+    const tree = mount(<BasicForm user={{ name: 'jared' }} />);
+    expect((tree.get(0).type as any).displayName).toBe('WithFormik(Form)');
   });
 
   describe('FormikHandlers', () => {
@@ -338,7 +338,7 @@ describe('withFormik()', () => {
       });
 
       describe('with validate (SYNC)', () => {
-        it('should call validate if present', () => {
+        it('should call validate if present', async () => {
           const validate = jest.fn().mockReturnValue({});
           const ValidateForm = withFormik<Props, Values, Values>({
             validate,
@@ -346,16 +346,14 @@ describe('withFormik()', () => {
             handleSubmit: noop,
           })(Form);
           const tree = mount(<ValidateForm user={{ name: 'jared' }} />);
-          tree
+          await tree
             .find(Form)
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
           expect(validate).toHaveBeenCalled();
         });
 
-        it('should submit the form if valid', () => {
+        it('should submit the form if valid', async () => {
           const handleSubmit = jest.fn();
           const ValidateForm = withFormik<Props, Values, Values>({
             validate: noop,
@@ -363,16 +361,14 @@ describe('withFormik()', () => {
             handleSubmit,
           })(Form);
           const tree = mount(<ValidateForm user={{ name: 'jared' }} />);
-          tree
+          await tree
             .find(Form)
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
           expect(handleSubmit).toHaveBeenCalled();
         });
 
-        it('should not submit the form if invalid', () => {
+        it('should not submit the form if invalid', async () => {
           const validate = jest.fn().mockReturnValue({ name: 'Error!' });
           const handleSubmit = jest.fn();
 
@@ -383,19 +379,17 @@ describe('withFormik()', () => {
           })(Form);
 
           const tree = mount(<ValidateForm user={{ name: '' }} />);
-          tree
+          await tree
             .find(Form)
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
           expect(validate).toHaveBeenCalled();
           expect(handleSubmit).not.toHaveBeenCalled();
         });
       });
 
       describe('with validate (ASYNC)', () => {
-        it('should call validate if present', () => {
+        it('should call validate if present', async () => {
           const validate = jest.fn(() => Promise.resolve({}));
           const ValidateForm = withFormik<Props, Values, Values>({
             validate,
@@ -403,12 +397,10 @@ describe('withFormik()', () => {
             handleSubmit: noop,
           })(Form);
           const tree = mount(<ValidateForm user={{ name: 'jared' }} />);
-          tree
+          await tree
             .find(Form)
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
           expect(validate).toHaveBeenCalled();
         });
 
@@ -453,20 +445,20 @@ describe('withFormik()', () => {
       });
 
       describe('with validationSchema (ASYNC)', () => {
-        it('should run validationSchema if present', () => {
+        it('should run validationSchema if present', async () => {
           const validate = jest.fn(() => Promise.resolve({}));
           const ValidateForm = FormFactory({ validationSchema: { validate } });
           const tree = mount(<ValidateForm user={{ name: 'jared' }} />);
-          tree
+
+          await tree
             .find(Form)
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
+
           expect(validate).toHaveBeenCalled();
         });
 
-        it('should call validationSchema if it is a function and present', () => {
+        it('should call validationSchema if it is a function and present', async () => {
           const validate = jest.fn(() => Promise.resolve({}));
           const ValidateForm = FormFactory({
             validationSchema: () => ({
@@ -474,12 +466,10 @@ describe('withFormik()', () => {
             }),
           });
           const tree = mount(<ValidateForm user={{ name: 'jared' }} />);
-          tree
+          await tree
             .find(Form)
-            .find('form')
-            .simulate('submit', {
-              preventDefault: noop,
-            });
+            .props()
+            .submitForm();
           expect(validate).toHaveBeenCalled();
         });
       });
