@@ -4,78 +4,111 @@ import replace from 'rollup-plugin-replace';
 import resolve from 'rollup-plugin-node-resolve';
 import sourceMaps from 'rollup-plugin-sourcemaps';
 import uglify from 'rollup-plugin-uglify';
+import pkg from './package.json';
 
-const shared = {
-  entry: `compiled/formik.js`,
-  sourceMap: true,
-  external: ['react', 'react-native'],
-  globals: {
-    react: 'React',
-    'react-native': 'ReactNative',
+const input = './compiled/index.js';
+const external = ['react', 'react-native'];
+
+const buildUmd = ({ env }) => ({
+  input,
+  external,
+  output: {
+    name: 'Formik',
+    format: 'umd',
+    sourcemap: true,
+    file:
+      env === 'production'
+        ? `./dist/formik.umd.${env}.js`
+        : `./dist/formik.umd.${env}.js`,
+    exports: 'named',
+    globals: {
+      react: 'React',
+      'react-native': 'ReactNative',
+    },
   },
-  exports: 'named',
-};
+
+  plugins: [
+    resolve(),
+    replace({
+      exclude: 'node_modules/**',
+      'process.env.NODE_ENV': JSON.stringify(env),
+    }),
+    commonjs({
+      include: /node_modules/,
+      namedExports: {
+        'node_modules/prop-types/index.js': [
+          'object',
+          'oneOfType',
+          'string',
+          'node',
+          'func',
+          'bool',
+          'element',
+        ],
+      },
+    }),
+    sourceMaps(),
+    env === 'production' && filesize(),
+    env === 'production' &&
+      uglify({
+        output: { comments: false },
+        compress: {
+          keep_infinity: true,
+          pure_getters: true,
+        },
+        warnings: true,
+        ecma: 5,
+        toplevel: false,
+      }),
+  ],
+});
+
+const buildCjs = ({ env }) => ({
+  input,
+  external: external.concat(Object.keys(pkg.dependencies)),
+  output: [
+    {
+      file: `./dist/${pkg.name}.cjs.${env}.js`,
+      format: 'cjs',
+      sourcemap: true,
+    },
+  ],
+  plugins: [
+    resolve(),
+    replace({
+      exclude: 'node_modules/**',
+      'process.env.NODE_ENV': JSON.stringify(env),
+    }),
+    sourceMaps(),
+    filesize(),
+  ],
+})
 
 export default [
-  Object.assign({}, shared, {
-    moduleName: 'Formik',
-    format: 'umd',
-    dest:
-      process.env.NODE_ENV === 'production'
-        ? './dist/formik.umd.min.js'
-        : './dist/formik.umd.js',
-    plugins: [
-      resolve(),
-      replace({
-        exclude: 'node_modules/**',
-        'process.env.NODE_ENV': JSON.stringify(
-          process.env.NODE_ENV || 'development'
-        ),
-      }),
-      resolve(),
-      commonjs({
-        include: /node_modules/,
-        namedExports: {
-          'node_modules/prop-types/index.js': [
-            'object',
-            'oneOfType',
-            'string',
-            'node',
-            'func',
-            'bool',
-            'element',
-          ],
-        },
-      }),
-      sourceMaps(),
-      process.env.NODE_ENV === 'production' && filesize(),
-      process.env.NODE_ENV === 'production' && uglify(),
-    ],
-  }),
-
-  Object.assign({}, shared, {
-    targets: [
-      { dest: 'dist/formik.es6.js', format: 'es' },
-      { dest: 'dist/formik.js', format: 'cjs' },
+  buildUmd({ env: 'production' }),
+  buildUmd({ env: 'development' }),
+  buildCjs({ env: 'production' }),
+  buildCjs({ env: 'development' }),
+  {
+    input,
+    external: external.concat(Object.keys(pkg.dependencies)),
+    output: [
+      {
+        file: pkg.module,
+        format: 'es',
+        sourcemap: true,
+      },
+      {
+        file: pkg.main,
+        format: 'cjs',
+        sourcemap: true,
+      },
     ],
     plugins: [
       resolve(),
-      commonjs({
-        include: /node_modules/,
-        namedExports: {
-          'node_modules/prop-types/index.js': [
-            'object',
-            'oneOfType',
-            'string',
-            'node',
-            'func',
-            'bool',
-            'element',
-          ],
-        },
-      }),
-      ,
+      
       sourceMaps(),
+      filesize(),
     ],
-  }),
+  },
 ];
