@@ -1,15 +1,12 @@
-import * as PropTypes from 'prop-types';
 import * as React from 'react';
-import { dlv } from './utils';
-
-import { FormikProps } from './formik';
-import { isFunction, isEmptyChildren } from './utils';
 import warning from 'warning';
-
-export type GenericFieldHTMLAttributes =
-  | React.InputHTMLAttributes<HTMLInputElement>
-  | React.SelectHTMLAttributes<HTMLSelectElement>
-  | React.TextareaHTMLAttributes<HTMLTextAreaElement>;
+import { connect } from './connect';
+import {
+  FormikProps,
+  GenericFieldHTMLAttributes,
+  FormikContext,
+} from './types';
+import { getIn, isEmptyChildren, isFunction } from './utils';
 
 /**
  * Note: These typings could be more restrictive, but then it would limit the
@@ -48,7 +45,10 @@ export interface FieldConfig {
   /**
    * Field component to render. Can either be a string like 'select' or a component.
    */
-  component?: string | React.ComponentType<FieldProps<any> | void>;
+  component?:
+    | string
+    | React.ComponentType<FieldProps<any>>
+    | React.ComponentType<void>;
 
   /**
    * Render prop (works like React router's <Route render={props =>} />)
@@ -58,7 +58,12 @@ export interface FieldConfig {
   /**
    * Children render function <Field name>{props => ...}</Field>)
    */
-  children?: ((props: FieldProps<any>) => React.ReactNode);
+  children?: ((props: FieldProps<any>) => React.ReactNode) | React.ReactNode;
+
+  /**
+   * Validate a single field value independently
+   */
+  validate?: ((value: any) => string | Promise<void> | undefined);
 
   /**
    * Field name
@@ -70,33 +75,27 @@ export interface FieldConfig {
 
   /** Field value */
   value?: any;
+
+  /** Inner ref */
+  innerRef?: (instance: any) => void;
 }
 
-export type FieldAttributes = GenericFieldHTMLAttributes & FieldConfig;
+export type FieldAttributes<T> = GenericFieldHTMLAttributes & FieldConfig & T;
 
 /**
  * Custom Field component for quickly hooking into Formik
  * context and wiring up forms.
  */
-
-export class Field<Props extends FieldAttributes = any> extends React.Component<
-  Props,
+class FieldInner<Props = {}, Values = {}> extends React.Component<
+  FieldAttributes<Props> & { formik: FormikContext<Values> },
   {}
 > {
-  static contextTypes = {
-    formik: PropTypes.object,
-  };
+  constructor(
+    props: FieldAttributes<Props> & { formik: FormikContext<Values> }
+  ) {
+    super(props);
 
-  static propTypes = {
-    name: PropTypes.string.isRequired,
-    component: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
-    render: PropTypes.func,
-    children: PropTypes.oneOfType([PropTypes.func, PropTypes.node]),
-  };
-
-  componentWillMount() {
-    const { render, children, component } = this.props;
-
+    const { render, children, component, formik } = props;
     warning(
       !(component && render),
       'You should not use <Field component> and <Field render> in the same <Field> component; <Field component> will be ignored'
@@ -111,23 +110,62 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
       !(render && children && !isEmptyChildren(children)),
       'You should not use <Field render> and <Field children> in the same <Field> component; <Field children> will be ignored'
     );
+
+    // Register the Field with the parent Formik. Parent will cycle through
+    // registered Field's validate fns right prior to submit
+    formik.registerField(props.name, {
+      validate: props.validate,
+    });
+  }
+
+  componentDidUpdate(
+    prevProps: FieldAttributes<Props> & { formik: FormikContext<Values> }
+  ) {
+    if (this.props.name !== prevProps.name) {
+      this.props.formik.unregisterField(prevProps.name);
+      this.props.formik.registerField(this.props.name, {
+        validate: this.props.validate,
+      });
+    }
+
+    if (this.props.validate !== prevProps.validate) {
+      this.props.formik.registerField(this.props.name, {
+        validate: this.props.validate,
+      });
+    }
+  }
+
+  componentWillUnmount() {
+    this.props.formik.unregisterField(this.props.name);
   }
 
   render() {
-    const { name, render, children, component = 'input', ...props } = this
-      .props as FieldConfig;
-
-    const { formik } = this.context;
+    const {
+      validate,
+      name,
+      render,
+      children,
+      component = 'input',
+      formik,
+      ...props
+    } = (this.props as FieldAttributes<Props> & {
+      formik: FormikContext<Values>;
+    }) as any;
+    const {
+      validate: _validate,
+      validationSchema: _validationSchema,
+      ...restOfFormik
+    } = formik;
     const field = {
       value:
         props.type === 'radio' || props.type === 'checkbox'
-          ? props.value
-          : dlv(formik.values, name),
+          ? props.value // React uses checked={} for these inputs
+          : getIn(formik.values, name),
       name,
       onChange: formik.handleChange,
       onBlur: formik.handleBlur,
     };
-    const bag = { field, form: formik };
+    const bag = { field, form: restOfFormik };
 
     if (render) {
       return (render as any)(bag);
@@ -138,9 +176,11 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
     }
 
     if (typeof component === 'string') {
+      const { innerRef, ...rest } = props;
       return React.createElement(component as any, {
+        ref: innerRef,
         ...field,
-        ...props,
+        ...rest,
         children,
       });
     }
@@ -152,3 +192,5 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
     });
   }
 }
+
+export const Field = connect<FieldAttributes<any>, any>(FieldInner);
