@@ -1,69 +1,41 @@
+import cloneDeep from 'lodash.clonedeep';
+import toPath from 'lodash.topath';
 import * as React from 'react';
 
-/** @private is the given object/value a promise? */
-export function isPromise(value: any): boolean {
-  if (value !== null && typeof value === 'object') {
-    return value && typeof value.then === 'function';
-  }
-
-  return false;
-}
-
-/** @private is running React Native?  */
-export const isReactNative =
-  typeof window !== 'undefined' &&
-  window.navigator &&
-  window.navigator.product &&
-  window.navigator.product === 'ReactNative';
-
-/** @private Returns values of an object in a new array */
-export function values<T>(obj: any): T[] {
-  const vals = [];
-  for (var key in obj) {
-    if (obj.hasOwnProperty(key)) {
-      vals.push(obj[key]);
-    }
-  }
-  return vals;
-}
-
-/** 
- * @private Deeply get a value from an object via it's dot path. 
- * See https://github.com/developit/dlv/blob/master/index.js
+/**
+ * Deeply get a value from an object via it's path.
  */
-export function dlv(
+export function getIn(
   obj: any,
   key: string | string[],
   def?: any,
   p: number = 0
 ) {
-  key = (key as string).split ? (key as string).split('.') : key;
-  while (obj && p < key.length) {
-    obj = obj[key[p++]];
+  const path = toPath(key);
+  while (obj && p < path.length) {
+    obj = obj[path[p++]];
   }
   return obj === undefined ? def : obj;
 }
 
-/** 
- * @private Deeply set a value from in object via it's dot path. 
- * See https://github.com/developit/linkstate
+/**
+ * Deeply set a value from in object via it's path.
+ * @see https://github.com/developit/linkstate
  */
-export function setDeep(path: string, value: any, obj: any): any {
+export function setIn(obj: any, path: string, value: any): any {
   let res: any = {};
   let resVal: any = res;
   let i = 0;
-  let pathArray = path.replace(/\]/g, '').split(/\.|\[/);
+  let pathArray = toPath(path);
 
   for (; i < pathArray.length - 1; i++) {
     const currentPath: string = pathArray[i];
-    let currentObj: any = obj[currentPath];
+    let currentObj: any = getIn(obj, pathArray.slice(0, i + 1));
 
     if (resVal[currentPath]) {
       resVal = resVal[currentPath];
     } else if (currentObj) {
-      resVal = resVal[currentPath] = Array.isArray(currentObj)
-        ? [...currentObj]
-        : { ...currentObj };
+      resVal = resVal[currentPath] = cloneDeep(currentObj);
     } else {
       const nextPath: string = pathArray[i + 1];
       resVal = resVal[currentPath] =
@@ -71,22 +43,103 @@ export function setDeep(path: string, value: any, obj: any): any {
     }
   }
 
-  resVal[pathArray[i]] = value;
-  return { ...obj, ...res };
+  if (value === undefined) {
+    delete resVal[pathArray[i]];
+  } else {
+    resVal[pathArray[i]] = value;
+  }
+
+  const result = { ...obj, ...res };
+
+  // If the path array has a single element, the loop did not run.
+  // Deleting on `resVal` had no effect in this scenario, so we delete on the result instead.
+  if (i === 0 && value === undefined) {
+    delete result[pathArray[i]];
+  }
+
+  return result;
 }
 
-/** @private is the given object a Function? */
-export const isFunction = (obj: any) => 'function' === typeof obj;
+/**
+ * Recursively a set the same value for all keys and arrays nested object, cloning
+ * @param object
+ * @param value
+ * @param visited
+ * @param response
+ */
+export function setNestedObjectValues<T>(
+  object: any,
+  value: any,
+  visited: any = new WeakMap(),
+  response: any = {}
+): T {
+  for (let k of Object.keys(object)) {
+    const val = object[k];
+    if (isObject(val)) {
+      if (!visited.get(val)) {
+        visited.set(val, true);
+        // In order to keep array values consistent for both dot path  and
+        // bracket syntax, we need to check if this is an array so that
+        // this will output  { friends: [true] } and not { friends: { "0": true } }
+        response[k] = Array.isArray(val) ? [] : {};
+        setNestedObjectValues(val, value, visited, response[k]);
+      }
+    } else {
+      response[k] = value;
+    }
+  }
 
+  return response;
+}
+
+// Assertions
+
+/** @private is the given object a Function? */
+export const isFunction = (obj: any): obj is Function =>
+  typeof obj === 'function';
 
 /** @private is the given object an Object? */
-export const isObject = (obj: any) => obj !== null && typeof obj === 'object';
+export const isObject = (obj: any): boolean =>
+  obj !== null && typeof obj === 'object';
+
+/** @private is the given object an integer? */
+export const isInteger = (obj: any): boolean =>
+  String(Math.floor(Number(obj))) === obj;
+
+/** @private is the given object a string? */
+export const isString = (obj: any): obj is string =>
+  Object.prototype.toString.call(obj) === '[object String]';
+
+/** @private is the given object a NaN? */
+export const isNaN = (obj: any): boolean => obj !== obj;
+
+/** @private Does a React component have exactly 0 children? */
+export const isEmptyChildren = (children: any): boolean =>
+  React.Children.count(children) === 0;
+
+/** @private is the given object/value a promise? */
+export const isPromise = (value: any): value is PromiseLike<any> =>
+  isObject(value) && isFunction(value.then);
 
 /**
- * @private is the given object an Integer? 
- * see https://stackoverflow.com/questions/10834796/validate-that-a-string-is-a-positive-integer
-*/
-export const isInteger = (obj: any) => String(Math.floor(Number(obj))) === obj;
-
-export const isEmptyChildren = (children: any) =>
-  React.Children.count(children) === 0;
+ * Same as document.activeElement but wraps in a try-catch block. In IE it is
+ * not safe to call document.activeElement if there is nothing focused.
+ *
+ * The activeElement will be null only if the document or document body is not
+ * yet defined.
+ *
+ * @param {?Document} doc Defaults to current document.
+ * @return {Element | null}
+ * @see https://github.com/facebook/fbjs/blob/master/packages/fbjs/src/core/dom/getActiveElement.js
+ */
+export function getActiveElement(doc?: Document): Element | null {
+  doc = doc || (typeof document !== 'undefined' ? document : undefined);
+  if (typeof doc === 'undefined') {
+    return null;
+  }
+  try {
+    return doc.activeElement || doc.body;
+  } catch (e) {
+    return doc.body;
+  }
+}
