@@ -1,61 +1,249 @@
-import { setDeep } from '../src/utils';
+import { setIn, setNestedObjectValues, isPromise } from '../src/utils';
 
-describe('helpers', () => {
-  describe('setDeep', () => {
+describe('utils', () => {
+  describe('setNestedObjectValues', () => {
+    it('sets value flat object', () => {
+      const obj = { x: 'y' };
+      expect(setNestedObjectValues(obj, true)).toEqual({ x: true });
+    });
+
+    it('sets value of nested object', () => {
+      const obj = {
+        x: {
+          nestedObject: {
+            z: 'thing',
+          },
+        },
+      };
+
+      const newObj = {
+        x: {
+          nestedObject: {
+            z: true,
+          },
+        },
+      };
+
+      expect(setNestedObjectValues(obj, true)).toEqual(newObj);
+    });
+
+    it('sets value of nested flat array items', () => {
+      const obj = {
+        x: {
+          nestedObject: {
+            z: 'thing',
+          },
+          nestedFlatArray: ['jared', 'ian'],
+        },
+      };
+
+      const newObj = {
+        x: {
+          nestedObject: {
+            z: true,
+          },
+          nestedFlatArray: [true, true],
+        },
+      };
+      expect(setNestedObjectValues(obj, true)).toEqual(newObj);
+    });
+
+    it('sets value of nested complex array items', () => {
+      const obj = {
+        x: {
+          nestedObject: {
+            z: 'thing',
+          },
+          nestedFlatArray: [
+            {
+              nestedObject: {
+                z: 'thing',
+              },
+            },
+            {
+              nestedObject: {
+                z: 'thing',
+              },
+            },
+          ],
+        },
+      };
+
+      const newObj = {
+        x: {
+          nestedObject: {
+            z: true,
+          },
+          nestedFlatArray: [
+            {
+              nestedObject: {
+                z: true,
+              },
+            },
+            {
+              nestedObject: {
+                z: true,
+              },
+            },
+          ],
+        },
+      };
+      expect(setNestedObjectValues(obj, true)).toEqual(newObj);
+    });
+    it('sets value of nested mixed array items', () => {
+      const obj = {
+        x: {
+          nestedObject: {
+            z: 'thing',
+          },
+          nestedFlatArray: [
+            {
+              nestedObject: {
+                z: 'thing',
+              },
+            },
+            'jared',
+          ],
+        },
+      };
+
+      const newObj = {
+        x: {
+          nestedObject: {
+            z: true,
+          },
+          nestedFlatArray: [
+            {
+              nestedObject: {
+                z: true,
+              },
+            },
+            true,
+          ],
+        },
+      };
+      expect(setNestedObjectValues(obj, true)).toEqual(newObj);
+    });
+  });
+
+  describe('setIn', () => {
     it('sets flat value', () => {
       const obj = { x: 'y' };
-      const newObj = setDeep('flat', 'value', obj);
+      const newObj = setIn(obj, 'flat', 'value');
       expect(obj).toEqual({ x: 'y' });
       expect(newObj).toEqual({ x: 'y', flat: 'value' });
     });
 
+    it('removes flat value', () => {
+      const obj = { x: 'y' };
+      const newObj = setIn(obj, 'x', undefined);
+      expect(obj).toEqual({ x: 'y' });
+      expect(newObj).toEqual({});
+      expect(newObj).not.toHaveProperty('x');
+    });
+
     it('sets nested value', () => {
       const obj = { x: 'y' };
-      const newObj = setDeep('nested.value', 'nested value', obj);
+      const newObj = setIn(obj, 'nested.value', 'nested value');
       expect(obj).toEqual({ x: 'y' });
       expect(newObj).toEqual({ x: 'y', nested: { value: 'nested value' } });
     });
 
     it('updates nested value', () => {
       const obj = { x: 'y', nested: { value: 'a' } };
-      const newObj = setDeep('nested.value', 'b', obj);
+      const newObj = setIn(obj, 'nested.value', 'b');
       expect(obj).toEqual({ x: 'y', nested: { value: 'a' } });
       expect(newObj).toEqual({ x: 'y', nested: { value: 'b' } });
     });
 
+    it('removes nested value', () => {
+      const obj = { x: 'y', nested: { value: 'a' } };
+      const newObj = setIn(obj, 'nested.value', undefined);
+      expect(obj).toEqual({ x: 'y', nested: { value: 'a' } });
+      expect(newObj).toEqual({ x: 'y', nested: {} });
+      expect(newObj.nested).not.toHaveProperty('value');
+    });
+
+    it('updates deep nested value', () => {
+      const obj = { x: 'y', twofoldly: { nested: { value: 'a' } } };
+      const newObj = setIn(obj, 'twofoldly.nested.value', 'b');
+      expect(obj.twofoldly.nested === newObj.twofoldly.nested).toEqual(false); // fails, same object still
+      expect(obj).toEqual({ x: 'y', twofoldly: { nested: { value: 'a' } } }); // fails, it's b here, too
+      expect(newObj).toEqual({ x: 'y', twofoldly: { nested: { value: 'b' } } }); // works ofc
+    });
+
+    it('removes deep nested value', () => {
+      const obj = { x: 'y', twofoldly: { nested: { value: 'a' } } };
+      const newObj = setIn(obj, 'twofoldly.nested.value', undefined);
+      expect(obj.twofoldly.nested === newObj.twofoldly.nested).toEqual(false);
+      expect(obj).toEqual({ x: 'y', twofoldly: { nested: { value: 'a' } } });
+      expect(newObj).toEqual({ x: 'y', twofoldly: { nested: {} } });
+      expect(newObj.twofoldly.nested).not.toHaveProperty('value');
+    });
+
     it('sets new array', () => {
       const obj = { x: 'y' };
-      const newObj = setDeep('nested.0', 'value', obj);
+      const newObj = setIn(obj, 'nested.0', 'value');
       expect(obj).toEqual({ x: 'y' });
       expect(newObj).toEqual({ x: 'y', nested: ['value'] });
     });
 
     it('updates nested array value', () => {
       const obj = { x: 'y', nested: ['a'] };
-      const newObj = setDeep('nested.0', 'b', obj);
+      const newObj = setIn(obj, 'nested[0]', 'b');
       expect(obj).toEqual({ x: 'y', nested: ['a'] });
       expect(newObj).toEqual({ x: 'y', nested: ['b'] });
     });
 
     it('adds new item to nested array', () => {
       const obj = { x: 'y', nested: ['a'] };
-      const newObj = setDeep('nested.1', 'b', obj);
+      const newObj = setIn(obj, 'nested.1', 'b');
       expect(obj).toEqual({ x: 'y', nested: ['a'] });
       expect(newObj).toEqual({ x: 'y', nested: ['a', 'b'] });
     });
 
     it('sticks to object with int key when defined', () => {
       const obj = { x: 'y', nested: { 0: 'a' } };
-      const newObj = setDeep('nested.0', 'b', obj);
+      const newObj = setIn(obj, 'nested.0', 'b');
       expect(obj).toEqual({ x: 'y', nested: { 0: 'a' } });
       expect(newObj).toEqual({ x: 'y', nested: { 0: 'b' } });
     });
 
     it('supports bracket path', () => {
       const obj = { x: 'y' };
-      const newObj = setDeep('nested[0]', 'value', obj);
+      const newObj = setIn(obj, 'nested[0]', 'value');
       expect(obj).toEqual({ x: 'y' });
       expect(newObj).toEqual({ x: 'y', nested: ['value'] });
+    });
+
+    it('supports path containing key of the object', () => {
+      const obj = { x: 'y' };
+      const newObj = setIn(obj, 'a.x.c', 'value');
+      expect(obj).toEqual({ x: 'y' });
+      expect(newObj).toEqual({ x: 'y', a: { x: { c: 'value' } } });
+    });
+  });
+
+  describe('isPromise', () => {
+    it('verifies that a value is a promise', () => {
+      const alwaysResolve = (resolve: Function) => resolve();
+      const promise = new Promise(alwaysResolve);
+      expect(isPromise(promise)).toEqual(true);
+    });
+
+    it('verifies that a value is not a promise', () => {
+      const emptyObject = {};
+      const identity = (i: any) => i;
+      const foo = 'foo';
+      const answerToLife = 42;
+
+      expect(isPromise(emptyObject)).toEqual(false);
+      expect(isPromise(identity)).toEqual(false);
+      expect(isPromise(foo)).toEqual(false);
+      expect(isPromise(answerToLife)).toEqual(false);
+
+      expect(isPromise(undefined)).toEqual(false);
+      expect(isPromise(null)).toEqual(false);
     });
   });
 });
