@@ -1,11 +1,11 @@
-import { validateYupSchema, yupToFormErrors } from '../src/formik';
+import { validateYupSchema, yupToFormErrors } from '../src';
 
 const Yup = require('yup');
 const schema = Yup.object().shape({
   name: Yup.string('Name must be a string').required('required'),
 });
 
-describe('helpers', () => {
+describe('Yup helpers', () => {
   describe('yupToFormErrors()', () => {
     it('should transform Yup ValidationErrors into an object', async () => {
       try {
